+++ scripts/accounts/accountInfo.js
@@ -1,0 +1,5 @@
+export const accountInfo = [
+  {email: "test@codeit.com", password: "codeit101"},
+  {email: "test2@codeit.com", password: "codeit202"},
+  {email: "test3@codeit.com", password: "codeit303"},
+]
