+++ signin.html
@@ -51,6 +51,6 @@
         </div>
       </div>
     </form>
-    <script defer src="./js/signin.js"></script>
+    <script type="module" src="./scripts/signin.js"></script>
   </body>
 </html>
