+++ signup.html
@@ -61,6 +61,6 @@
           </div>
         </div>
     </form>
-    <script defer src="./scripts/signup.js"></script>
+    <script type="module" src="./scripts/signup.js"></script>
   </body>
 </html>
